05practicerauth/pyproject.toml
[project]
name = "practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.3",
    "email-validator>=2.2.0",
    "fastapi[all]>=0.116.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.24",
    "uvicorn[standard]>=0.35.0",
]
